# API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bestfit": {
            "post": {
                "tags": [
                    "Memory Allocation"
                ],
                "summary": "Simulate Best Fit Memory Allocation Algorithm",
                "description": "This API simulates the Best Fit memory allocation algorithm by accepting block sizes and process sizes to allocate memory efficiently.",
                "operationId": "33f796d5ed22955cdbe86203c36919a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "Block",
                                    "Process"
                                ],
                                "properties": {
                                    "Algorithm": {
                                        "description": "Type of algorithm (e.g., bestfit)",
                                        "type": "string",
                                        "example": "bestfit"
                                    },
                                    "Block": {
                                        "description": "Sizes of memory blocks (space-separated)",
                                        "type": "string",
                                        "example": "100 500 200 300 600"
                                    },
                                    "Process": {
                                        "description": "Sizes of processes (space-separated)",
                                        "type": "string",
                                        "example": "212 417 112 426"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation with memory allocation results"
                    },
                    "400": {
                        "description": "Invalid request due to incorrect input"
                    }
                }
            }
        },
        "/api/firstfit": {
            "post": {
                "tags": [
                    "Memory Allocation"
                ],
                "summary": "Simulate First Fit Memory Allocation Algorithm",
                "description": "This API simulates the First Fit memory allocation algorithm by accepting block sizes and process sizes to allocate memory efficiently.",
                "operationId": "ce506d84623b5f0c7408f089828dd626",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "Block",
                                    "Process"
                                ],
                                "properties": {
                                    "Algorithm": {
                                        "description": "Type of algorithm (e.g., firstfit)",
                                        "type": "string",
                                        "example": "firstfit"
                                    },
                                    "Block": {
                                        "description": "Sizes of memory blocks (space-separated)",
                                        "type": "string",
                                        "example": "100 500 200 300 600"
                                    },
                                    "Process": {
                                        "description": "Sizes of processes (space-separated)",
                                        "type": "string",
                                        "example": "212 417 112 426"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation with memory allocation results"
                    },
                    "400": {
                        "description": "Invalid request due to incorrect input"
                    }
                }
            }
        },
        "/api/worstfit": {
            "post": {
                "tags": [
                    "Memory Allocation"
                ],
                "summary": "Simulate First Fit Memory Allocation Algorithm",
                "description": "This API simulates the Worst Fit memory allocation algorithm by accepting block sizes and process sizes to allocate memory efficiently.",
                "operationId": "c8d4f736dab8575b4e8316f4272de71e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "Block",
                                    "Process"
                                ],
                                "properties": {
                                    "Algorithm": {
                                        "description": "Type of algorithm (e.g., worstfit)",
                                        "type": "string",
                                        "example": "worstfit"
                                    },
                                    "Block": {
                                        "description": "Sizes of memory blocks (space-separated)",
                                        "type": "string",
                                        "example": "100 500 200 300 600"
                                    },
                                    "Process": {
                                        "description": "Sizes of processes (space-separated)",
                                        "type": "string",
                                        "example": "212 417 112 426"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation with memory allocation results"
                    },
                    "400": {
                        "description": "Invalid request due to incorrect input"
                    }
                }
            }
        },
        "/api/nextfit": {
            "post": {
                "tags": [
                    "Memory Allocation"
                ],
                "summary": "Simulate Next Fit Memory Allocation Algorithm",
                "description": "This API simulates the Next Fit memory allocation algorithm by accepting block sizes and process sizes to allocate memory efficiently.",
                "operationId": "ff4dc1546e04720cd66d034405b07eb9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "Block",
                                    "Process"
                                ],
                                "properties": {
                                    "Algorithm": {
                                        "description": "Type of algorithm (e.g., nextfit)",
                                        "type": "string",
                                        "example": "nextfit"
                                    },
                                    "Block": {
                                        "description": "Sizes of memory blocks (space-separated)",
                                        "type": "string",
                                        "example": "100 500 200 300 600"
                                    },
                                    "Process": {
                                        "description": "Sizes of processes (space-separated)",
                                        "type": "string",
                                        "example": "212 417 112 426"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation with memory allocation results"
                    },
                    "400": {
                        "description": "Invalid request due to incorrect input"
                    }
                }
            }
        },
...
    },
...
}